+++ package.json
@@ -9,27 +9,27 @@
     "astro": "astro"
   },
   "dependencies": {
-    "@astrojs/mdx": "^2.3.0",
-    "@swup/astro": "^1.4.0",
+    "@astrojs/mdx": "^3.1.6",
+    "@swup/astro": "^1.4.1",
     "@tailwindcss/line-clamp": "^0.4.4",
-    "@tailwindcss/typography": "^0.5.12",
-    "astro": "^4.6.0",
+    "@tailwindcss/typography": "^0.5.15",
+    "astro": "^4.15.7",
     "astro-font": "^0.0.77",
     "astro-imagetools": "^0.9.0",
     "astro-meta": "^0.0.2",
-    "astro-seo": "^0.8.3",
-    "cssnano": "^6.1.2",
-    "dayjs": "^1.11.10",
+    "astro-seo": "^0.8.4",
+    "cssnano": "^7.0.6",
+    "dayjs": "^1.11.13",
     "mdast-util-to-string": "^4.0.0",
     "reading-time": "^1.5.0",
-    "tailwindcss": "^3.4.3"
+    "tailwindcss": "^3.4.12"
   },
   "devDependencies": {
-    "@astrojs/tailwind": "^5.1.0",
-    "autoprefixer": "^10.4.19",
-    "postcss": "^8.4.38",
-    "prettier": "^3.2.5",
-    "prettier-plugin-tailwindcss": "^0.5.13",
-    "vite": "^5.2.8"
+    "@astrojs/tailwind": "^5.1.1",
+    "autoprefixer": "^10.4.20",
+    "postcss": "^8.4.47",
+    "prettier": "^3.3.3",
+    "prettier-plugin-tailwindcss": "^0.6.6",
+    "vite": "^5.4.6"
   }
 }
